split out game code into module
