+++ backend/core/go.mod
@@ -1,18 +1,21 @@
 module main
 
-go 1.24
+go 1.24.0
 
-require github.com/julienschmidt/httprouter v1.3.0
+replace sheepshead => ../common/sheepshead
 
 require (
 	cloud.google.com/go/firestore v1.15.0
 	firebase.google.com/go/v4 v4.14.1
 	github.com/ably/ably-go v1.2.21
 	github.com/joho/godotenv v1.5.1
+	github.com/julienschmidt/httprouter v1.3.0
 	github.com/rs/cors v1.10.1
 	github.com/sio/coolname v0.1.0
 	github.com/sirupsen/logrus v1.4.2
 	github.com/stretchr/testify v1.10.0
+	google.golang.org/grpc v1.62.1
+	sheepshead v0.0.0-00010101000000-000000000000
 )
 
 require (
@@ -56,7 +59,6 @@ require (
 	google.golang.org/genproto v0.0.0-20240213162025-012b6fc9bca9 // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20240314234333-6e1732d8331c // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20240311132316-a219d84964c2 // indirect
-	google.golang.org/grpc v1.62.1 // indirect
 	google.golang.org/protobuf v1.33.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	nhooyr.io/websocket v1.8.17 // indirect
